fixes types
