+++ lib/idx/proceed.ts
@@ -14,26 +14,11 @@
 import { 
   OktaAuthInterface,
   IdxTransaction,
+  ProceedOptions,
 } from '../types';
 import { run } from './run';
-import { AuthenticationOptions } from './authenticate';
-import {
-  EnrollPollValues as EnrollPollOptions,
-  SelectEnrollmentChannelValues as SelectEnrollmentChannelOptions
-} from './remediators';
-import { RegistrationOptions } from './register';
-import { PasswordRecoveryOptions } from './recoverPassword';
-import { AccountUnlockOptions } from './unlockAccount';
 import { getSavedTransactionMeta } from './transactionMeta';
 import { AuthSdkError } from '../errors';
-
-export type ProceedOptions = AuthenticationOptions
-  & RegistrationOptions
-  & PasswordRecoveryOptions
-  & AccountUnlockOptions
-  & EnrollPollOptions
-  & SelectEnrollmentChannelOptions
-  & { step?: string };
 
 export function canProceed(authClient: OktaAuthInterface, options: ProceedOptions = {}): boolean {
   const meta = getSavedTransactionMeta(authClient, options);
