+++ lib/idx/remediate.ts
@@ -14,8 +14,7 @@
 /* eslint-disable max-statements, max-depth, complexity */
 import { AuthSdkError } from '../errors';
 import { RemediationValues } from './remediators';
-import { FlowIdentifier, RemediationResponse } from './types';
-import { RemediationFlow } from './flow';
+import { RemediateOptions, RemediationResponse } from './types';
 import { 
   IdxResponse,
   IdxActionParams, 
@@ -35,13 +34,6 @@ export interface RemediateActionWithOptionalParams {
 }
 
 export type RemediateAction = string | RemediateActionWithOptionalParams;
-export interface RemediateOptions {
-  remediators?: RemediationFlow;
-  actions?: RemediateAction[];
-  flow?: FlowIdentifier;
-  step?: string;
-  shouldProceedWithEmailAuthenticator?: boolean; // will be removed in next major version
-}
 
 
 function getActionFromValues(values: RemediationValues, idxResponse: IdxResponse): string | undefined {
